backend/requirements.txt
bcrypt==4.1.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==3.0.0
Flask-JWT-Extended==4.5.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
PyJWT==2.8.0
pymongo==4.6.0
python-dotenv==1.0.0
Werkzeug==3.0.1
